controller cleanup; routes cleanup; removed package from auto discovery

+++ src/app/Http/Controllers/AddressesController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
 use LaravelEnso\AddressesManager\app\Enums\StreetTypes;
+use LaravelEnso\AddressesManager\app\Exceptions\AddressException;
 use LaravelEnso\AddressesManager\App\Http\Requests\ValidateAddressRequest;
 use LaravelEnso\Core\app\Exceptions\EnsoException;
 use LaravelEnso\FormBuilder\app\Classes\FormBuilder;
@@ -14,17 +15,13 @@ use LaravelEnso\RoAddresses\app\Models\County;
 
 class AddressesController extends Controller
 {
-    public function index()
-    {
-        return view('laravel-enso/addressesmanager::index');
-    }
 
     public function store(ValidateAddressRequest $request, string $type, int $id)
     {
         $address = new Address($request->all());
         $address->addressable_id = $id;
         $address->addressable_type = config('addresses.addressables.'.$type);
-        $address->is_default = $this->isTheFirst($address) ?: false;
+        $address->is_default = $this->isTheFirst($address);
 
         $address->save();
 
@@ -36,8 +33,7 @@ class AddressesController extends Controller
 
     public function update(ValidateAddressRequest $request, Address $address)
     {
-        $address->fill($request->all());
-        $address->save();
+        $address->update($request->all());
 
         return [
             'message' => __('The Changes have been saved!'),
@@ -55,14 +51,7 @@ class AddressesController extends Controller
     public function setDefault(Address $address)
     {
         DB::transaction(function () use ($address) {
-
-            //first set all addresses as not default
-            $address->addressable->addresses()->where('is_default', true)->get()
-                ->each(function (Address $item) {
-                    $item->is_default = false;
-                    $item->save();
-                });
-
+            $this->unsetDefaultAddress($address);
             $address->is_default = true;
             $address->save();
         });
@@ -81,6 +70,9 @@ class AddressesController extends Controller
      */
     public function destroy(Address $address)
     {
+        if ($address->is_default) {
+            throw new AddressException(__('The default address cannot be deleted'));
+        }
         $address->delete();
 
         return [
@@ -89,7 +81,7 @@ class AddressesController extends Controller
         ];
     }
 
-    public function getEditForm(Address $address)
+    public function edit(Address $address)
     {
         $editForm = (new FormBuilder($this->getFormPath(), $address))
             ->setTitle('Edit')
@@ -102,7 +94,7 @@ class AddressesController extends Controller
         return $editForm;
     }
 
-    public function getCreateForm(Request $request)
+    public function create(Request $request)
     {
         $postUrl = sprintf('/addresses/%s/%s',
             $request->get('addressable_type'), $request->get('addressable_id'));
@@ -127,7 +119,7 @@ class AddressesController extends Controller
     {
         $addressable = $this->getAddressable();
 
-        return $addressable->addresses()->get();
+        return $addressable->addresses()->orderBy('is_default', 'desc')->get();
     }
 
     /**
@@ -170,6 +162,18 @@ class AddressesController extends Controller
         }
 
         return __DIR__.'/../../Forms/addresses/address.json';
+    }
+
+    private function unsetDefaultAddress(Address $address)
+    {
+        $defaultAddress = $address->addressable->addresses()
+            ->whereIsDefault(true)
+            ->first();
+
+        if (!is_null($defaultAddress)) {
+            $defaultAddress->is_default = false;
+            $defaultAddress->save();
+        }
     }
 
     private function isTheFirst(Address $address)
